trip/requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
pytz==2018.6
